IcLog rework to fix lag + extra changes (#4)
The IC Log was reworked so that it reduces the amount of work that is performed to render messages. This massively helps improve performance, even to the point of cutting down a full second in accepting messages when the IC Log is large.

A new `config.ini` parameter was also introduced: `chatlog_newline` (which defaults to False). If False, IC log messages will be rendered as `<name>: <message>`; if True, IC log messages will be rendered as `<name>
<message>

`
(three lines). 

Finally, `scroll_type` in `config.ini` was changed to `chatlog_scrolldown`. `scroll_type = down` is equivalent to `chatlog_scrolldown = true` and `scroll_type = up` is equivalent to `chatlog_scrolldown = false`.

Co-authored-by: Chrezm <TheChrezm@gmail.com>