Removed APNG plugin loader
Qt loads the APNG plugin by itself given the location is within the folder the client is ran from (`imageformats/<qapng plug-in>`).

The QPluginLoader is thus no longer necessary and is now redundant, thus safe to remove.

TLDR; It's no longer needed and we required a long-worded explanation as to why.