Added audio-device selection
DRAudioEngine has been updated to automatically pick a device and can be provided a favorite device.

If a favorite device has been designed, DRAudioEngine will attempt to transfert all currently playing sounds to the favorite device of choice.

Otherwise, it will attempt to pick the designated audio device chosen by the OS.

Finally, If there is no designated audio device, it will pick the first one available.

Finally, if there are no devices available the client will simply prevent audio from playing and will report the issue in the console.

DRAudioEngine will always attempt to have an audio device active. As long as one is enabled within the OS, it will attempt its best to use it. Failing this, sounds will simply be disabled and won't be able to play until an appropriate device is available.

**Favorite Device**
* To select a favorite device simply switch to the wanted device in the dropdown menu in AOConfigPanel
**NOTE: The dropdown will always reflect the currently used device. If clicked on the dropdown, the favorite device will be highlighted in green to indicate it's the favorite device.

If clicking outside of the dropdown, it will simply close the dropdown list of choices and will continue showing the current device.**